requirements.txt
torch
face_recognition
annoy
imutils
opencv-python
matplotlib
numpy
tqdm
pillow
onnx2torch
requests